change getArmRotations functions
So they get the distance from a target number of rotations rather than just returning the number of rotations.
They still use the average of both the motor rotations. That seems like a recipe for disaster, but I don't think relying on just one of the motors is good either.
- Alex G.