Climb: use Rev Neo relative encoder instead of duty cycle encoder
The duty cycle encoder only reports a value from [0, 1)--this can be
biased, but is still restricted to a range of 1.0--and is also behind a
gear assembly, so encoder turns multiple times for one revolution of
arm.

Use relative encoder instead, which has wider range of motion.  Use duty
cycle encoder to set initial position of relative encoder on startup.

- Justin C