Climb: stop motor when arm swings with gravity assist
Imagine the arm at 0 degress (vertical).  It rotates backward 90 degress
to point behind the robot.  Now set the target to 90 degrees ahead of
the robot.  From -90 deg to 0 deg, gravity will do most of the work, and
motor does not need to run.  From 0 deg to 90 deg the arm must be pushed
against gravity, and the motor is required.

Due to friction, add a +/- 6.0 unit window around zero where the motor
kicks on to overcome friction.

- Justin C