Fix(Climb): do not overdrive into high bar indefinitely
The subsequent stage will rotate the outer arms to vertical in a race
group and the timing there works such that we can discard this
malfunctioning rotation stage.

By "timing works", I mean the robot climbs fast enough that outer arms
rotate so the hooks are above the high bar instead of below.

- Justin C.