Climb: bump acceptable velocity error to 0.75 and reset velocity error
Ran into a wierd bug.  The first time we run the climb sequence
(high or traversal) the inner arms quickly clear the latches and the
sequence continues.  If we don't restart the robot code and run the
sequence again, the inner arms overshoot and oscillate and never reach
their target position.

Adjust acceptable error and reset velocity error in an attempt to curb
this behavior.

Note: it's not detrimental to competition since the robot code restarts
before every match.

- Justin C.