Climb: add setpoint to clear previous bar when rotating to vertical
The back arms can hit the previous bar.  Pull the arms in closer to the
robot so this doesn't happen.

Previous tests relied (incorrectly) on the substantial swing of the
robot to miss the previous bar.  The slow pids reduce the swing of the
robot and now the arms hit the previous bar.

- Justin C.